# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Savewell savings tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsCategory } from './analyticsCategory';
import type { AnalyticsDay } from './analyticsDay';

export interface Analytics {
  range: string;
  /** @minimum 0 */
  totalPaise: number;
  /** @minimum 0 */
  averagePaise: number;
  /** @nullable */
  bestDay: Date | null;
  byCategory: AnalyticsCategory[];
  byDay: AnalyticsDay[];
}