# lib/api-zod/src/generated/types/analyticsCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Savewell savings tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsCategory {
  name: string;
  icon: string;
  /** @minimum 0 */
  totalPaise: number;
}