# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Savewell savings tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 8
     */
  icon: string;
}