# lib/api-zod/src/generated/types/savingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Savewell savings tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SavingUpdate {
  /** @minimum 1 */
  amountPaise?: number;
  /** @minimum 1 */
  categoryId?: number;
  /** @nullable */
  goalId?: number | null;
  date?: Date;
  /**
     * @maxLength 240
     * @nullable
     */
  note?: string | null;
  isGoalLinked?: boolean;
}